+++ src/AllowAll.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Access;
+
+/**
+ * Allow all access.
+ */
+final class AllowAll implements AccessCheckerInterface
+{
+    public function userHasPermission($userId, string $permissionName, array $parameters = []): bool
+    {
+        return true;
+    }
+}